chore: Refactor InfraService and ddress remaining PR comments.
