+++ errors/InitializationError.ts
@@ -1,0 +1,8 @@
+export default class InitializationError extends Error {
+  constructor(message: string) {
+    super()
+    Error.captureStackTrace(this, this.constructor)
+    this.name = this.constructor.name
+    this.message = message || "An error occurred during initialization"
+  }
+}
